ENH: add support for win64 for visual studio 2005 ide and nmake, also fix warnings produced by building for win64

+++ Source/CTest/cmCTestHandlerCommand.cxx
@@ -38,7 +38,7 @@ cmCTestHandlerCommand::cmCTestHandlerCommand()
 bool cmCTestHandlerCommand::InitialPass(
   std::vector<std::string> const& args)
 {
-  if ( !this->ProcessArguments(args, this->Last, &*this->Arguments.begin(),
+  if ( !this->ProcessArguments(args, (unsigned int)this->Last, &*this->Arguments.begin(),
       this->Values) )
     {
     return false;
